+++ src/Encamina.Enmarcha.SemanticKernel/EphemeralMemoryStoreHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Encamina.Enmarcha.SemanticKernel.Abstractions;
+using Encamina.Enmarcha.SemanticKernel.Abstractions.Events;
 using Encamina.Enmarcha.SemanticKernel.Options;
 
 using Microsoft.Extensions.Logging;
@@ -21,8 +22,20 @@ internal sealed class EphemeralMemoryStoreHandler : MemoryStoreHandlerBase
     /// <param name="memoryManager">A valid instance of <see cref="IMemoryManager"/> that manages the memory store handled by this instance.</param>
     /// <param name="sessionManagementOptions">Configuration options for this memory store handler.</param>
     /// <param name="logger">A logger for this memory store handler.</param>
+    [Obsolete("This constructor is obsolete and will be removed in a future version. Use the Encamina.Enmarcha.SemanticKernel.EphemeralMemoryStoreHandler.EphemeralMemoryStoreHandler(IMemoryStoreExtender memoryStoreExtender, IOptionsMonitor<EphemeralMemoryStoreHandlerOptions> sessionManagementOptions, ILogger<EphemeralMemoryStoreHandler> logger) constructor instead.", false)]
     public EphemeralMemoryStoreHandler(IMemoryManager memoryManager, IOptionsMonitor<EphemeralMemoryStoreHandlerOptions> sessionManagementOptions, ILogger<EphemeralMemoryStoreHandler> logger)
-        : base(memoryManager)
+        : this(memoryManager as IMemoryStoreExtender, sessionManagementOptions, logger)
+    {
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="EphemeralMemoryStoreHandler"/> class.
+    /// </summary>
+    /// <param name="memoryStoreExtender">A valid instance of <see cref="IMemoryStoreExtender"/> that extends the memory store handled by this instance.</param>
+    /// <param name="sessionManagementOptions">Configuration options for this memory store handler.</param>
+    /// <param name="logger">A logger for this memory store handler.</param>
+    public EphemeralMemoryStoreHandler(IMemoryStoreExtender memoryStoreExtender, IOptionsMonitor<EphemeralMemoryStoreHandlerOptions> sessionManagementOptions, ILogger<EphemeralMemoryStoreHandler> logger)
+        : base(memoryStoreExtender)
     {
         this.logger = logger;
 
@@ -52,7 +65,8 @@ internal sealed class EphemeralMemoryStoreHandler : MemoryStoreHandlerBase
                 foreach (var memoryStoreInfo in MemoryStoreCollectionInfo.Where(i => i.Value.LastAccessUtc < date).ToList())
                 {
                     MemoryStoreCollectionInfo.Remove(memoryStoreInfo.Key);
-                    await MemoryManager.MemoryStore.DeleteCollectionAsync(memoryStoreInfo.Value.CollectionName, cancellationToken);
+                    await MemoryStoreExtender.MemoryStore.DeleteCollectionAsync(memoryStoreInfo.Value.CollectionName, cancellationToken);
+                    MemoryStoreExtender.RaiseMemoryStoreEvent(new() { EventType = MemoryStoreEventTypes.DeleteCollection, CollectionName = memoryStoreInfo.Value.CollectionName });
                 }
 
                 Thread.Sleep(TimeSpan.FromMinutes(options.InactivePollingTimeMinutes));